Save Day
Worked on Save shit all day. What a nightmare, I hate ResourceSaver and everything about it. For some reason, my new method of saving causes 0.5 sec lag, and the file is 2mb and miles long. isn't that rly cool and awesome?

also, started Wallet. probably did something else too, who knows

this is the worst kind of progress. it was all necessary, but it felt awful the whole time