Final Save Day!
This will be the final MAJOR Save Day. Woo-hoo!!! Everything loads correctly!!! No doubt there will be some bugs that slipped through the cracks, but I will fix those as I notice them. Over time, this save/load system will be perfect! And it's all done in one scene without creating new LOREDs or Upgrades or anything else. (It deletes all Wishes and creates new Wishes in their place.)

There will be over 300 Upgrades by the end of Stage 3, so it's a good thing that loading them is so easy and fast.

All I have to do now to "finalize" it is get it to save to file, and save to clipboard. Currently it is only saving to a test variable in-game. But I am certain this won't take more than 10 minutes. :D

So anyway, after this, I want to get back to implementing the Wallet, which is the game's resource viewer/manager. From there, I'll keep adding main Wishes until the whole of Stage 2 is caught up and working. I'm sure that getting resetting (prestiging) to work will take at least a day, but because of all the work I did for Saves, I'm estimating that it will be pretty painless (famous last words).

bye-bye