Limit Break!
Behind-the-scenes code is done! Thanks to signals, I didn't have to go around my scripts adding a bunch of lines like "limit_break.add_xp(amount_produced)". All I did was connect the currency's "increased_by_lored" signal to the add_xp function. It was already established that whenever a LORED increases a currency (as opposed to a Wish), it will emit that signal.
So, yeah, the LB script is  only 80 lines (for now)!

Tomorrow I will create the HUD for it. Here's my notes for that:

1. in UpgradeEffect for LIMIT_BREAK, the apply_method should be up.limit_break.enabled.set_true, and the opposite for remove_method

2. create HUD for LB. xp bar and w/e. in its script, connect limit_break.enabled.changed to a func which shows/hides the HUD.

3. in the _ready for LB HUD, for its progress bar, do attach_attribute(up.limit_break.xp)

I'm pretty sure that will be it! Done! And I already know the saving works because I mistakenly let it save the values and I thought it was buggy! The Limit Break XP was starting off at e28.00 and I was like "WHY IS THIS BROKEN" but actually it was working perfectly!

Feels great to have things go smoothly instead of crashing into a wall every 5 seconds. That's how it felt when working on the game in the old version which didn't take advantage of signals.