Dynamic Upgrades will be in Sidebar!
Wondering where to put Limit Break, I was thinking of creating a top bar, but decided to put it in the sidebar! This will keep the screen clean & organized. Just throw all the shit in the sidebar. Heck yeah.

I worked a lot today (at my real job)! Not enough time for LORED :pensive: