Limit Break Done!
Now I will begin adding some QoL shit related to dynamic Upgrades, like being able to pick one to display on the main screen. For that, I would need to create a new widget, but that will be ok and easy.

I also need to hide the widgets if their Upgrade is not purchased (even if you already purchased it in the past).