LOREDs can level up multiple times at once!
Changed all Upgrades that affect LORED Cost in any way.

Something is wrong! There is a huge balancing issue where LOREDs can sprint to level 1,000 quickly! I don't know why it would happen in this version and not the live version. The Upgrade effects are very similar.

Well, that's annoying. And once again I'm trying to redesign the LORED UI. I just don't like how it looks, and surely it is hard to understand for new players. Ugh... dis is hard.