Right Sidebar Progress
All Upgrades except 2 are tested and CONFIRMED to work. If they are found to not work by any player, that will simply be the last straw.

There seems to be something funky with autobuyers where after a bit they stop purchasing. I'll have to find out why!

-

I worked on LORED a lot today, but it doesn't FEEL like I got much done. Some days are just like that, I guess! Everything I did was necessary and took as long as it needed to take. But it FELT like I didn't make as much progress as I would have liked to make. But dat okay.

okej cya